packages/core/pyproject.toml
[project]
name = "agent-factory-core"
version = "0.1.0"
description = "Shared foundation for agent-factory studios"
requires-python = ">=3.11"
dependencies = [
    "deepagents",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.1",
    "langgraph>=0.4",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "colorthief>=0.2",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "sse-starlette>=2.0",
    "jinja2>=3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
